Convert character QTable to numeric if possible
* BasicTable/AsBasicTable will check if the input
data has all numeric or NA entries and convert to
numeric in this case.  If any character entries are
present, the BasicTable is returned as with all
character entries with a warning.

* Certain statistics in Q produce character output, in which
case the other statistics in the Qtable will also be numeric.
Need to ensure that we convert to numeric if we are dropping
to only use the first statistic and it happens to be numeric.

Addresses: DS-1471