Add support for lists in RemoveRowsAndOrColumns
* Input data in RemoveRowsAndOrColumns can now
be a list of matrices and data.frames

* Input data in RemoveByName can not be a list
of vectors. Each element of x is searched for names
matching the removal names, if a match is found elements
are removed from each element of x.  If there are no
matches, an attempt will be made to remove elements
of x by name

* Implemented to allow for cleaner code in flipChart::PrepareData
in the case of the input data being a list of user-supplied tables
Addresses: DS-1471, DS-1530